How we can connect two services
How we can connect two services: - build up communication between two services

1.IAM users access key or secreate key
2.With the help of Role we can connect two services.

We see 2. With the help of "Role"

1. create instance >> select instancce >> security >> ModifyIAMrole >> create role >> (AWS service, Use cases:- EC2, Use case with S3 service >>> add permission:-S3FullAccess)
  >>Attach role and connect or launch instance "without putting access key or Seccreate access key"

#aws s3 ls								........because we connect 2 services using "Role"

Default usernames of all Linux ec2 instance while login mobaxterm or any third party application:-

For Amazon Linux 2 or the Amazon Linux AMI, the user name is ec2-user.

For a CentOS AMI, the user name is centos or ec2-user.

For a Debian AMI, the user name is admin.

For a Fedora AMI, the user name is fedora or ec2-user.

For a RHEL AMI, the user name is ec2-user or root.

For a SUSE AMI, the user name is ec2-user or root.

For an Ubuntu AMI, the user name is ubuntu.

For an Oracle AMI, the user name is ec2-user.

For a Bitnami AMI, the user name is bitnami.